requirements.txt
streamlit
PyMuPDF
requests
Pillow
python-dotenv
tiktoken
nltk
python-docx
celery
redis
azure-storage-blob
scikit-learn
PyPDF2
python-pptx
python-docx
azure-storage-blob
